+++ jest.config.js
@@ -5,6 +5,10 @@ process.env.TZ = 'UTC';
 module.exports = {
   // Jest configuration provided by Grafana scaffolding
   ...require('./.config/jest.config'),
+  moduleNameMapper: {
+    ...require('./.config/jest.config').moduleNameMapper,
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
 
   /**
    * Reset mocks implementation between tests