+++ src/components/FormPanel/FormPanel.tsx
@@ -19,7 +19,8 @@ import {
   RefreshEvent,
   toDataQueryResponse,
 } from '@grafana/runtime';
-import { Alert, Button, ButtonGroup, ConfirmModal, usePanelContext, useStyles2, useTheme2 } from '@grafana/ui';
+import { Alert, Button, ConfirmModal, usePanelContext, useStyles2, useTheme2 } from '@grafana/ui';
+import { CustomButtonsRow } from 'components/CustomButtonsRow';
 import { isEqual } from 'lodash';
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
@@ -80,6 +81,9 @@ export const FormPanel: React.FC<Props> = ({
   replaceVariables,
   data,
 }) => {
+  /**
+   * State
+   */
   const [loading, setLoading] = useState<LoadingMode>(LoadingMode.INITIAL);
   const [error, setError] = useState('');
   const [title, setTitle] = useState('');
@@ -222,7 +226,7 @@ export const FormPanel: React.FC<Props> = ({
       response?: FetchResponse | Response | DataQueryResponse | null;
       initialRequest?: () => Promise<void>;
       currentElements?: LocalFormElement[];
-    }) => {
+    }): Promise<void> => {
       if (!code) {
         return;
       }
@@ -928,6 +932,7 @@ export const FormPanel: React.FC<Props> = ({
               initial={initial}
               section={null}
               replaceVariables={replaceVariables}
+              executeCustomCode={executeCustomCode}
             />
           </div>
         ) : (
@@ -940,13 +945,20 @@ export const FormPanel: React.FC<Props> = ({
             replaceVariables={replaceVariables}
             sectionsExpandedState={sectionsExpandedState}
             onChangeSectionExpandedState={onChangeSectionExpandedState}
+            executeCustomCode={executeCustomCode}
           />
         )}
       </div>
-      <ButtonGroup className={cx(styles.button[options.buttonGroup.orientation])}>
+      <div className={cx(styles.buttons, styles.buttonsPosition[options.buttonGroup.orientation])}>
+        <CustomButtonsRow
+          data={data}
+          executeCustomCode={executeCustomCode}
+          initial={initial}
+          elements={elements}
+          replaceVariables={replaceVariables}
+        />
         {options.updateEnabled !== UpdateEnabledMode.DISABLED && (
           <Button
-            className={cx(styles.margin)}
             variant={getButtonVariant(options.submit.variant)}
             icon={loading === LoadingMode.UPDATE ? 'fa fa-spinner' : options.submit.icon}
             title={title}
@@ -977,7 +989,6 @@ export const FormPanel: React.FC<Props> = ({
 
         {options.reset.variant !== ButtonVariant.HIDDEN && (
           <Button
-            className={cx(styles.margin)}
             variant={getButtonVariant(options.reset.variant)}
             icon={loading === LoadingMode.RESET ? 'fa fa-spinner' : options.reset.icon}
             style={
@@ -1008,7 +1019,6 @@ export const FormPanel: React.FC<Props> = ({
 
         {options.saveDefault.variant !== ButtonVariant.HIDDEN && canSaveDefaultValues && (
           <Button
-            className={cx(styles.margin)}
             variant={getButtonVariant(options.saveDefault.variant)}
             icon={options.saveDefault.icon}
             disabled={!!loading || !saveDefaultEnabled}
@@ -1020,7 +1030,7 @@ export const FormPanel: React.FC<Props> = ({
             {options.saveDefault.text}
           </Button>
         )}
-      </ButtonGroup>
+      </div>
 
       {error && (
         <Alert data-testid={TEST_IDS.panel.errorMessage} severity="error" title="Request">